+++ Final project/app/src/main/java/it/polito/mad/playgroundsreservations/reservations/ShowReservationFragment.kt
@@ -123,7 +123,7 @@ class ShowReservationFragment: Fragment(R.layout.show_reservation_fragment) {
                 val ratingBarValue = viewModel.getRatingsByPlaygroundIdFragment(myPlayground.id)
                 ratingBarValue.observe(viewLifecycleOwner) { ratingPlaygroundsList ->
                     if (ratingPlaygroundsList.isEmpty()){
-                        ratingBar.rating = (0.0).toFloat()
+                        ratingBar.rating = 0.0f
                     } else {
                         ratingPlaygroundsList.forEach { r ->
                             totalRating += r?.rating!!
